SCC-2221 Add check-in card data
This fetches the associated check-in card data (consisting of an array of "check-in boxes") for the holding record being parsed. This is separately fetched from the Sierra database and is used to display the most recently acquired items in a serial.
