+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 ### Added
 - Initial commit containing basic structure and location parser
 - Added supporting documentation and unit test suite
+- Added integration with check-in card data
