+++ lib/record_manager.rb
@@ -1,3 +1,5 @@
+require_relative './field_parser'
+
 class RecordManager
     def initialize record
         @record = record
@@ -27,7 +29,31 @@ class RecordManager
         # Extract all holdings strings into the holdings field
         _get_h_fields
 
-        # TODO Incorporate Check-In Card data to this object
+        # Incorporate Check-In Card data to this object
+        _get_check_in_card
+    end
+
+    def _get_check_in_card
+        $logger.debug "Fetching check-in card data for holding ##{@record['id']}"
+        
+        begin
+            check_in_card_uri = URI("#{ENV['PLATFORM_API_BASE_URL']}/holdings/check-in-cards?holding_id=#{@record['id']}")
+            response = Net::HTTP.get_response(check_in_card_uri)
+        rescue Exception => e
+            $logger.error "Failed to load check-in card data", { :status => e.message }
+            raise RecordError.new("Could not load check-in-card data")
+        end
+        
+        # Confirm that a valid response was received
+        unless response.code.to_i == 200 
+            $logger.error "Unable to load check-in card data with status #{response.code}", { :status => response.body }
+            raise RecordError.new("Unable to load check-in card data with status #{response.code}")
+        end
+
+        # Parse response into an object
+        check_in_data = JSON.parse(response.body)
+        $logger.debug check_in_data
+        @record['check_in_card'] = check_in_data
     end
 
     def _get_h_fields
@@ -92,79 +118,4 @@ class RecordManager
     end
 end
 
-
-class ParsedField
-    attr_reader :string_rep
-
-    @@enumeration_codes = "abcdef"
-    @@chronology_codes = "ijkl"
-
-    def initialize(h_field, y_field)
-        @h_field = h_field
-        @y_field = y_field
-        @string_rep = ''
-    end
-
-    def generate_string_representation
-        enumeration = _generate_enumeration
-        chronology = _generate_chronology
-        @string_rep += enumeration.length > 0 ? enumeration : ''
-        if chronology.length > 0
-            chronology = enumeration.length > 0 ? " (#{chronology})" : chronology
-            @string_rep += chronology
-        end
-    end
-
-    private
-
-    def _generate_enumeration
-        @@enumeration_codes.split('').map { |c| @h_field.include?(c) ? "#{@y_field[c]} #{@h_field[c]}" : nil }.compact.join(', ')
-    end
-
-    def _generate_chronology
-        date_component = DateComponent.new
-        components = @@chronology_codes.split('').map do |c|
-            if @h_field.include? c 
-                date_component.set_field(@y_field[c].tr('()', ''), @h_field[c])
-            end
-        end
-
-        date_component.create_str
-
-        date_component.date_str
-    end
-
-    class DateComponent
-        attr_reader :date_str
-
-        @@dash_regex = /(?:[\-]{2,}|[\-]$)/
-
-        def initialize
-            @start_year = nil
-            @end_year = nil
-            @start_month = nil
-            @end_month = nil
-            @start_day = nil
-            @end_day = nil
-
-            @date_str = ''
-        end
-
-        def set_field(component, value) 
-            value_arr = value.split('-')
-            self.instance_variable_set("@start_#{component}", value_arr[0])
-            self.instance_variable_set("@end_#{component}", value_arr[1] ? value_arr[1] : value_arr[0])
-        end
-
-        def create_str
-            start_str = "#{@start_year}-#{@start_month}-#{@start_day}".gsub(@@dash_regex, '')
-            end_str = "#{@end_year}-#{@end_month}-#{@end_day}".gsub(@@dash_regex, '')
-
-            if start_str == end_str
-                @date_str = start_str
-            else
-                @date_str = "#{start_str}/#{end_str}"
-            end
-        end
-    end
-end
+class RecordError < StandardError; end
